+++ volmdlr/surfaces.py
@@ -3606,6 +3606,21 @@ class SphericalSurface3D(PeriodicalSurface):
         # Hidden Attributes
         self._bbox = None
 
+    def _circle_generatrixes(self, number_circles: int):
+        """
+        Gets the sphere circle generatrixes.
+
+        :param number_circles: number of circles to be created.
+        :return: List of Circle 3D.
+        """
+        circles = []
+        i_frame = volmdlr.Frame3D(self.frame.origin, self.frame.v, self.frame.w, self.frame.u)
+        for theta in npy.linspace(0, volmdlr.TWO_PI / 2, number_circles):
+            i_frame_ = i_frame.rotation(self.frame.origin, self.frame.w, theta)
+            circle = curves.Circle3D(i_frame_, self.radius)
+            circles.append(circle)
+        return circles
+
     @property
     def bounding_box(self):
         """Bounding Box for Spherical Surface 3D."""
@@ -4206,14 +4221,8 @@ class SphericalSurface3D(PeriodicalSurface):
             ax = fig.add_subplot(111, projection='3d')
 
         self.frame.plot(ax=ax, ratio=self.radius)
-        for i in range(20):
-            theta = i / 20. * volmdlr.TWO_PI
-            t_points = []
-            for j in range(20):
-                phi = j / 20. * volmdlr.TWO_PI
-                t_points.append(self.point2d_to_3d(volmdlr.Point2D(theta, phi)))
-            ax = wires.ClosedPolygon3D(t_points).plot(ax=ax, edge_style=edge_style)
-
+        for circle in self._circle_generatrixes(50):
+            circle.plot(ax, edge_style)
         return ax
 
     def rectangular_cut(self, theta1, theta2, phi1, phi2, name=''):